done roles and permissions setup

+++ app/Http/Controllers/Api/Company/CompanyController.php
@@ -35,8 +35,13 @@ class CompanyController extends Controller
      */
     public function index()
     {
-        return CompanyResource::collection(
-            Company::all());
+        if(auth()->user()->can('show-all-company')){
+            return CompanyResource::collection(
+                Company::all());
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
+
     }
 
     /**
@@ -47,12 +52,17 @@ class CompanyController extends Controller
      */
     public function store(StoreCompanyRequest $request)
     {
-        $company = Company::create($request->validated());
+        if(auth()->user()->can('add-company')){
+            $company = Company::create($request->validated());
 
-        return $this->success([
-            'data' => new CompanyResource($company),
-            'message' => 'Company created successfully',
-        ], 201);
+            return $this->success([
+                'data' => new CompanyResource($company),
+                'message' => 'Company created successfully',
+            ], 201);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
+
     }
 
     /**
@@ -63,10 +73,15 @@ class CompanyController extends Controller
      */
     public function show(Company $company)
     {
+        if(auth()->user()->can('show-company')){
+            return new CompanyResource(
+                Company::findOrFail($company->id)
+            );
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
         // return $this->isNotAuthorized($company) ? $this->isNotAuthorized($company) : new CompanyResource($company);
-        return new CompanyResource(
-            Company::findOrFail($company->id)
-        );
+
     }
 
     /**
@@ -74,8 +89,13 @@ class CompanyController extends Controller
      */
     public function update(Request $request, Company $company)
     {
-        $company->update($request->all());
-        return new CompanyResource($company);
+        if(auth()->user()->can('update-company')){
+            $company->update($request->all());
+            return new CompanyResource($company);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
+
     }
 
     /**
@@ -87,7 +107,12 @@ class CompanyController extends Controller
 //            return $this->isNotAuthorized($company);
 //        }
 
-        $company->delete();
-        return $this->success(null, 'Company deleted successfully', 200);
+        if(auth()->user()->can('delete-company')){
+            $company->delete();
+            return $this->success(null, 'Company deleted successfully', 200);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
+
     }
 }
